backend/requirements.txt
# ─── CrewAI (existing) ────────────────────────────────
crewai[tools,google-genai]>=0.100.0
python-dotenv>=1.0.0
supabase>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# ─── LangGraph Integration (new) ─────────────────────
langgraph>=0.4.0
langchain>=0.3.0
langchain-google-genai>=2.1.0
langchain-community>=0.3.0
langsmith>=0.3.0
requests>=2.31.0
